added regression tests
